+++ test-foundry/factory/PoolFactory.t.sol
@@ -38,7 +38,7 @@ contract PoolFactoryTest is DeployTest {
             address quote,
             address oracleAdapter,
             UD60x18 strike,
-            uint64 maturity,
+            uint256 maturity,
             bool isCallPool
         ) = IPool(pool).getPoolSettings();
